Make sure clipping is on if line extension is applied to pscoast -N
See issue #1388 for context.  The problem was that we implemented the fix that allows fat lines to properly be clipped at map boundaries (by extending the line outside the map and turn clipping on) but for this case the clipping was not activated, allow the extension line to be seen.  Now the dedicion to clip matches the code that allows line extension plus it applies to -W -N and -I as long as -M is not issued.