Use SHA256 hash for remote files and rename variables (#1627)
The hash codes on the server are now SHA256 instead of MD5.  This PR renames local variables in gmt_remote.? that used md5 or MD5 in their names to more general naems using "hash".  ALso, the file with the hash codes are now called gmt_hash_server.txt instead of gmt_md5_server.txt but I added a symbolic link so the old name will still find the hash table.