Add support for auto-legend if legend not called
Now, when legend is called and there is no input file and a hidden legend file is found, we use it and then delete it.  THus, when gmt inset end, gmt subplot end, or gmt end come along and find a legend file it means legend was not called explicitly.  If so, we add a call to legend with a few default settings.