Set clip path for map insets by default (#1760)
* WIP add -N to gmt inset begin

By default, we turn on clipping for the inset box.  As for pstext, we add -N to turn that off.  However, since gmt inset end does not know that we need a solution embedded in the PostScript.

* Define PSL_inset_clip

Need to be set whether clipping is on or not.

* Add documentation

Explaining -N.

* Check if panel is defined

Otherwise dont access it.