Make sure psconvert maintains correct PS level
WHen we finish the half-baked plot we open the file with PSL_fopen and close implicitly by calling PLS_endplot.  However the latter decrement the call_level while it was never incremented in the first place since PSL_plotbegin was not called in this situation.  Only affects the auto-legend stuff in #1749.