Let -Jz|Z be supplied from history if -p and a 3-D -R is in effect
See issue #1265 for context.  This PR stores the most recent vertical projection ( z or Z) in a Z history item so that we later can determine if we should supply -Jz or -JZ from history.  We do this if a 3-D plot is selected, which we identify if -p is set and the region has three dimensions (5 slashes in the -R string).