Let grdtrack -E+c continue track if next line is a continuation of previous (#1885)
* WIP Let grdtrack -E+c continue track if next line is a continuation of previous

If we need several intermediate point to define a line then we want distance to increase continuously and we want the result to be a single segment.  Currently we get individual line segments with distance reset to zero for each segment.  The +c modifier will tied these segments together once implemented.

* Prepare the documentation

Explain the **+c** modifier.

* Implement first draft hookup

* Closer to working

Need to fix parsing of modifiers each time?

* Improve the docs

* Let +c be parsed before lines since it applies to all