Prevent libcurl from downloading nonexistant files
If there is a typo in an earth_relief_xxy name then we still try to download from the remote server, wasting time and getting confusing error messages from the server.  This PR introduces a list of valid names and prescreens so taht we can nip this problem in the bud on the client side.