Create an implied or explicit dir in psconvert
See #1983.  This PR splits this problem into two solutions: 1. If -Ddir is used in pscovnert then we create that directory if it does not exist.  2. If a registered figure has a directory as part of its filename then we split that into a dir and file when calling psconvert and use -D and F.
Closes #1983.