Do not have default paint for items not selected (#1995)
* Do not have default paint for items not selected

This is in relation to the bug in #1982.  The question is one of defaults: pscoast starts by setting these:

		gmt_init_fill (GMT, &C->S.fill, 1.0, 1.0, 1.0);		/* Default Ocean color = white */
	C->C.fill[LAKE] = C->C.fill[RIVER] = C->S.fill;		/* Default Lake/Riverlake color = Ocean color */
	gmt_init_fill (GMT, &C->G.fill, 0.0, 0.0, 0.0);		/* Default Land color = black */

When a user typically sets land color via -G then the ocean fill is set to white even if -S is not used.  However, if -S is set but not -G (like in the users case), the land default is set to black.  Those are the black areas that show up in the bug.
THe solution is to set areas not requested to no_fill (-1,-1,-1) so they are never painted.  Doing so fixes the bug, but a few examples fail (2) like GMT_az_equidistant which sets land color only.  So ocean is unpainted (and hence paper white) but a lake taks on the same color as the ocean so it used to be white but now (in my fix is not painted).

* If a special case then set ocean to white

Forgot that for some -JE cases we must paint the ocean white, and same for lakes.