Modernize ancient /fontsize in psvelo
This one was missed during the reformation.  In other tools (e.g., psmeca) we replaced this slash-delimited zie-only setting with the modifier +ffont which not only controls font size but all attributes of the font.  Backwards compatible, and default set to 9p as before.  Tests seems to all run just as well.