Modernize ancient /fontsize in psvelo (#2049)
* Modernize ancient /fontsize in psvelo

This one was missed during the reformation.  In other tools (e.g., psmeca) we replaced this slash-delimited zie-only setting with the modifier +ffont which not only controls font size but all attributes of the font.  Backwards compatible, and default set to 9p as before.  Tests seems to all run just as well.

* eliminate spaces

* Clarify scale unit

* Remove nonesitant -F reference and update fill specs

Too old explanation of patterns.

* Update src/geodesy/psvelo.c

Co-Authored-By: Dongdong Tian <seisman.info@gmail.com>

* Fix slash same way as in meca