Giving a minus to indicate transparent pixels is lame
The GMT4 syntax for transparent patters had syntax like BredF-.  We replaced that with separate +b and +f modifier that take the require color.  However, we let - still signify transparency, yet there is no reason not to just assume no color given means transparency.  To be backwards compatible we accept either - or nothing, but the docs have been updated to say the default is transparent.