When postscript output is selected, psconvert uses the wrong file (#2141)
psconvert reads the original postscript file and writes it to a temp file, possibly adding more PostScript code for cropping or painting.  However, under modern mode, postscript is a choice, and then we simply copy the hidden postscript file to the output name.  However, the code copies the original file, not the updated temporary file.