Reset font settings when building PostScript macros (#2163)
Like we did for the PSL_plot_completion PostScript macro used by subplots to add the panel label after the plotting, we must do the same for the PSL_movie_completion macro used to label a movie panel: We may compute the font and font encoding to make the macro, but since the macro is not called until the end we must undo any memory of having encoded this font so that when the beginning of the plot needs those fonts it will have to encode them at that time.  Otherwise, when movie labeles were selected the degree symbols for basemaps disappeared (or similar).