Add modifier +e to region extender parsing and use functions
See #2165 and #2179 for context.  This PR adds +e to join +r and +R as modifiers of regions.  The +e is similar to +r except it ensure that the boundary moves at least 0.25 times the increment, thus always giving some space between data and boundary while still yielding boundaries in multiples of the increments.  Furthermore, the parsing of modifiers and the extending of the region are now encoded in functions so that we were able to add these modifiers to -I in gmtinfo.