Let psimage work with 1-bit images again (#2186)
Since GDAL returns 1-bit images as 8-bits with a colormap that has 2 entries, we needed to redo the logic in psimage.  IT is now wordking, and enw test bitimage.sh shows various ways to replace the fore or background color with a specified color, or to make one of them transparent.  Note the case of setting the background (black) to transparent yields a white block but if converted to PNG it shows the black pixels as transparent.