Add new share/tools script to get remote files (#2219)
* Add new sahre/tools script to get remote files

It is convenient for users to download all the cache and server DEMs prior to trips where internet connectivity may be poor or lacking.  gmt_getremote.sh downloads all the files not already on the system.

* Update CMakeLists.txt

Needed to add this script.  WOuld be better to have some wildcards here though

* Address offline usage

* Update share/tools/gmt_getremote.sh

Co-Authored-By: Dongdong Tian <seisman.info@gmail.com>

* Update gmt_getremote.sh

* Update share/tools/gmt_getremote.sh

Co-Authored-By: Dongdong Tian <seisman.info@gmail.com>