Allow both -i and -o to specify an open-ended list to end of record
If a user wants to shuffle the order of cols 0 and 1 but otherwise read the entire record, they have to list everything that follows the -i1,0 list to ensure all numerical columns are added.  She also needs to add ,t to ensure the trailing text is added.  Likewise for output.  This PR allows open-ended ranges to be given, such as

-i1,0,4:,t [read col 1, 0, then 4 to the end or record, plus trailing text]
-o3:,t1  [Write cols 3 to end of record, plus 2nd word of trailing text]