Let the "Edit on GitHub" button work for all branches
For the documentation of the master branch, the "Edit on GitHub" should
be linked to the master branch, but now it's linked to the non-existent
6.1 branch.

In this PR, the conf.py script is modified to check the environmental
variable BUILD_SOURCEBRANCHNAME predefined by Azure Pipelines. So we
can get the branch name from the environmental variable and set the
correct branch name to link. For locally built documentations, the
variable is not defined, then the current major and minor version is
used.