Supply -Baf when colorbar plots a stretched CPT (#2311)
* Supply -Baf when colorbar plots a stretched CPT

Tools like grdiimage may produce a stretched CPT taht fits the range of a grid exactly.  In those cases we cannot let psscale use the CPT levels as annotation and tick intervals but must override via -Baf.  This PR does so by adding a keyword STRETCHED to such CPTs as they are written and then we watch for that word upon reading.  CLoses issue #2307.

* Update colorbar_common.rst_

* Update src/gmt_support.c

Co-Authored-By: Dongdong Tian <seisman.info@gmail.com>

* Update src/gmt_support.c

Co-Authored-By: Dongdong Tian <seisman.info@gmail.com>

* Update doc/rst/source/colorbar_common.rst_

Co-Authored-By: Dongdong Tian <seisman.info@gmail.com>

* Better keyword name

* Must use -Baf for all stretched CPTs.

Co-authored-by: Dongdong Tian <seisman.info@gmail.com>