Bending over backwards to detect global NASA grids
Some NASA grids clearly has global dimensions but very sloppy coordinate arrays and increments tricked GMT into thinking it was not global.  I have added an extra test for exact 360 and 180 ranges and if those are found we screw those bad coordinates and go with the global nature of things.  Addrresses #2329.