Address issues with grdinfo -D option
See #2364.  THe intent was for -D to print the region strings, -D -C or -Cn to print the w e s n columns, and -D -Ct to print the 4 coumns plus the region string as trailing text.  I have updated the docs to make this clear plus fixed the code.  Closes #2364.