grdinfo would not print x and y unit names unless containing the word degree
There was a special provision to look for " [degrees" as part of the x and y unit strings and if so we just stripped off that part from the name.  However, if we had a Cartesian grid then that strstr returned NULL and we did nothng, including not printing out the original unit name.