Fix contour wrapping in the y-direction (#2399)
See #2389 for context.  When contours are converted to (x,y) locations on the map we then see if there is any wrapping across the map due to the periodic nature of longitudes.  However, for the transverse Mercator projection such wrapping can only happen in the y-direction and we had no check for that.  I have added a new check that deals with both of these potential wrapping lines.  In the process I discovered a bug or two in the old x-only wrapping checker but apparently it did not cause any actual damage.  Anyway, now fixed, but I added a basic example to the grdcontour tests to show this is fixed.  Closes issue #2389.