psl_fix_utf8: out_string needs one more byte (#2400)
In the function `psl_fix_utf8`, out_string is a new string with some UTF8
codes fixed. out_string should be the same length as in_string. When
allocating memory for `out_string`, we should give it one more byte for '\0'.
Otherwise, we may lose one byte in the `in_string` or have a
un-terminated `in_string` (no '\0') at the end of the function.