Report correct file when complaining about record (#2406)
See #2268 for context.  The issue is that when a ShapeFile is given then we convert it via ogr2ogr and make a temporary file, then read that file instead.  However, if an error occurs (such as here when we expect 3 columns but only 2 were present) the error message referred to a previous filename since we had not updated that variable.  Now we show more explicitly that we are reading a temp file created by ogr2ogr form the original input file.  Closes #2268.