Do not round off w/e/sn when -D -I are used for tiling grids
The -D option in grdinfo is used in conjunction with -I to report the subregions of tiles of given size.  However, the code got caught by a desire to round off the w/e/s/n region before tiling, leading to tiles exceeding the range of the grid.  Now fixed.