Fix cmake CMP0026 (#855)
CMP0026 (introduced in CMake 3.0) disallows use of the LOCATION property for build targets, because the LOCATION property is not fully determined until generate-time.

This means we cannot get LOCATION property of the libraries.

THis PR sets CMP0026 policy to new (i.e. disallows use of the LOCATION property). The filenames of the libraries are now determined by OUTPUT_NAME and PREFIX in the target property, and also CMAKE_SHARED_LIBRARY_SUFFIX, CMAKE_STATIC_LIBRARY_SUFFIX, CMAKE_SHARED_MODULE_SUFFIX settings.