Verious fixes
Progress bar 1 now plots in TR corner (not listening to +j) and only the background circle is shown - not the wedge yet.  Does respond to +Gfill though, and the wedge angles are passed correctly but no wedge - looks more like a spider set of lines, so probably not calling plot_symbol correctly.