doc: The "Edit On GitHub" icon should link to the master branch by default
For the "Edit On GitHub" icon on the documentations, we need to specify
a branch to link.

Previously, it's by default linked to the branch "version", i.e.
for master branch, the GMT version is 6.1, which doesn't exist.

This PR fixes the issue, by linking the icon to the master branch, which
is always available.