polar: Fix several bugs and improve documentations (#2461)
The two examples in the manpage don't work due to two bugs.

The first one, when takeoff-angle>90, symbol_size2 isn't changed, i.e.
symbol_size2=0, and symbols aren't plotted. It doesn't make sense.

The second one: I don't know much about HYPO71 format, but from the
codes and the example, the last column should be something like `ipu0`,
and the phase polarity is determined by its third character. Thus, when
read the last column, we should use `%s` instead of `%c`.

The original PS `seis_09.ps` is also updated, due to the bug 1.
When takeoff-angle=90, the symbol_size2 is set to 0.
The new PS is consistent with the output of GMT 4.5.18.