pspolar: Improve -T syntax and keep backwards-compatibility (#2462)
The -T syntax (-T<angle>/<form>/<justify>/<fontsize> is very old-style.

This PR improves the -T syntax via modifiers. The new -T syntax is:

    -T+a<angle>+f<font>+j<just>+o<dx>/<dy>

which is very similar to the psmeca -S option.

---

Changes in this PR:

- New -T syntax while keeping backwards-compatibility
- A new test to check the new -T syntax
- Move the two pspolar tests into directory `test/pspolar`
---

Although I said the new -T option is backwards-compatible, actually it's
not. The old -T syntax is still supported, but the default justification
and offset is different. The old codes use `PSL_ML` as the default
justification. Here I use `PSL_TC` instead, to make it consistent with
the **psmeca -S** option. I think such small incompatibility is OK to
most users.