Pass ints to*_setfill not bool
Most of our GMT_setfill and PSL_setfill calls passed the last arguments as true or false (directly or via a bool variable).  However, these functions expect ints, and not even an unsigned int.  For instance GMT_setfill can accept outline = -2 as a valid argument. While there seem to be no bugs in GMT regarding this mi of ints and bools, I have replaced true|false with 1|0 and changed arguments from bool to int in most places.  The confusion stems from old GMT 4 where we had #define TRUE 1 and #define FALSE 0 which where ints and these became true and false in GMT 5 but of course these are of type bool, not signed int.