Let movie -S provide PostScript layers directly as well as via scripts (#2473)
Some users may already have a PostScript layer that they wish to use as a backtround or foreground in a movie.  It is thus convenient that movie -S can handle that case.  It simply inserts the users specified PostScript file instead of the one created by the script.
Also noticed that the documentation for -G was incomplete.