Update how CPT hinges are handled (#2416)
* Update how CPT hinges are done

Separate between soft and hard hinges.  Hard hinges are data driven like those for bathymetry or magnetics, while soft hinges are ignored unless +h[hinge] is given.

* Updates so far - in progress

* Add warning that +h cannot affect hard hinges

* Update scripts and PS to show short and hard hinges

* Update GMT_tut_17.ps

* Update docs

Explain soft and hard hinges

* Add include files explaining CPT syntax for creation and use

* Polish cpt docs

* Fix handing of soft hinge assignment

* Improve hinge documentationand CPT

* Update list of CPTs to indicate hard or soft hinge

* Update world.cpt

* Update haxby.cpt

* Turn wysiwig, haxby, no_green, and world into continuous CPTs

Update a few PS files that changed.

* Restrict CPT resampling to continuous CPTs

We ban the resampling of discrete CPTs to another set of nodes.  Only stretching of the range is allowed.

* Allow colorlists to pass as discrete input CPT

* Strengthening handling of color lists and categorical tables

* Update std_opts_classic.rst

typo

* Add new CPT section on color-lists

Adds a new script figure as well

* Improve colorlist figure and discussion

* Update features.rst

* Check hinge and warn if outside data range (and ignore)

* Add new CPT test for hinges

* Update common_SYN_OPTs.rst_

* Update common_SYN_OPTs.rst_

* Deal with distcalc ref

* Better reporting for hinge violations in CPTs

* Allow slow or zhigh to equal hinge

Not fully working yet.  If hinge == zlow or hinge == zhigh we need to eliminate one entry in the resulting CPT (not done yet).

* Update doc/rst/source/cookbook/features.rst

Co-Authored-By: Dongdong Tian <seisman.info@gmail.com>

* Update src/gmt_support.c

Co-Authored-By: Dongdong Tian <seisman.info@gmail.com>

* Update gmt_api.c

* Improve GMT_colorlist.sh script to show options

* Update gmt_api.c

* Fix CPT stretching with hinges

Needed to update tut 19 and its PS.

* Fix discrete resampling of CPTs with hinges

Trim the CPT if needed before resampling.

* Address soft hinge and fix -T parsing check

I had labeled a CPT with soft hinge as a hard hinge if +h was given, but it should stay a soft hinge internally.  Also, add validation of possible modifiers for parse_array.

* Fix print message in validate_cpt function

* Leave user zmin/zmax selection as is

* Add new test, update cpthinges to new behavior

* Avoid freeing same array twice

Co-authored-by: Dongdong Tian <seisman.info@gmail.com>