Move ex03 to SI
See #2471.  I am also removing unnecessary -X -Y settings that were used in the old PostScript days to position the plot reasonably centered on the page.  Since we are producing PDF by design and ripping to PNG for the web page we should keep the examples as uncluttered of old positioning as possible.
Give there is so many palces where we nudge using points I decided it made sense here to give symbol sice in points (an even 5p) but it is OK to use cm as well and we can and should mix that up (but not use inches).