Extract tick length as size of fronts in legends
The initial screening of symbols in legend tries to determine the largest symbol size so it can use that as a default later.  However, front gets size/gap/tlength so to parse that we need to isolate tlength since it is closest to the notion of symbol size (size here is actually front length).  This is all done to avoid messages like
pslegend [WARNING]: 2i/0.6i/0.25 not a valid number and may not be decoded properly.
when running.