Address complications for filenames with spaces in psconvert and show
See #2646 for background.  The solution involved several steps:

The filename argument to psconvert -F needed to be de-quoted.  This is necessary because when we build the gs command we surround the final output name with quotes.
GMT_Create_Options need to consider both single and double quotes when dealing with single text string commands that contain filenames in quotes.
the gmt end show code also needed to be mindful of the presence of quotes in the file name.

With these changes, GMT handles filenames with spaces provided given in quotes.  Hopefully this addresses all of #2646 but that needs to be determined.