Compute require colatitude when setting a pole a point for small circle
project can draw a small circle, but if you specify the projection via a pole and a point C, it is complicated to draw the small circle through C because you do not know its colatitude without computing it first separately (e.g., with mapproject).  New modifier +c to -G will automatically supply the required colatitude to achive the desired small circle.