Ensure mapproject -G is parsed correctly (#2673)
See question on the forum.  The -G parser failed to parse various cases. A new checker was coded up that determins if the -G argument is of the new or deprecated type and calls the corresponding parser.  The usage message had deprecated syntax on how to switch calculation method.  PR affected several files as I moved gmtlib_count_char from an internal library function to being accessible by the modules and hence changed name to gmt_count_char.