Reduce useless messages from gmtconnect (#2792)
* Reduce useless messages from gmtconnect

gmtconnect does a first pass to determine which segments will be joined so that we can build a text string with these connection for the segment header.  The second pass then oes the connections.  THere is a smalll difference in how the two passes operate that makes the number of passes sometimes repoert different numbers of connections.  This is more of a debug message than anything else and hopefully this can be compleely fixed soon.  THus, I have moved this mesage from warning to debug so taht id does not bother people.

* spaces to tabs