Check if variable "file" is not NULL
GCC reports many warnings like the one below.
```
../src/gmt_support.c: In function ‘support_init_custom_symbol’:
../src/gmt_support.c:4678:27: warning: the address of ‘file’ will always evaluate as ‘true’ [-Waddress]
  if (gmt_M_file_is_cache (file)) /* Must be a cache file */
                           ^
../src/gmt_macros.h:200:36: note: in definition of macro ‘gmt_M_file_is_cache’
 #define gmt_M_file_is_cache(file) (file && file[0] == '@' && strncmp (file, "@GMTAPI@-", 9U))
```

When variable `file` is declared as a pointer like `char *file = NULL`,
gcc doesn't raise an warning. But when variable `file` is declared as an
string array like `char file[PATH_MAX] = {""}`, gcc reports the warning,
since the address of string `file` is alawys evaluated to be "true".

One solution is checking if `file != NULL`, which works for both string
array and pointer.