Optimize images before making a release or deploying to GitHub
In this PR, I added a new optional cmake target "optimize_images", which
compresses all PNG images used in the documentation.
Currently, the compression can reduce the images from 33 Mb to 18 Mb.

"optimize_images" is also added as a dependency of "gmt_release".
So, when we make release tarballs, the images in the `doc_release` directory are
always compressed.

Note that, the target "optimize_images" needs "pngquant" installed.