More robust parsing of fonts (#2889)
I found that if I gave Times-Italic,10p the parser got confused (it is expecting size,name.  We do quite a bit of checking on the argument order but this particular case was not tested.  I added one more check and now the above is robustly parsed.  No tests affected.