Allow oblique Mercator maps have a vertical oblique Equator
THrought GMT history, the oblique Equator has always been projected to be the horizontal x-axis.  However, for features with more N-S extend than E-W, it makes more sense to align the plot so that the oblique Equator is the positive y-axis.  This requires a new modifer to -JO (+v) to make it so.