Add modifier for a vertical oblique Equator in -JO (#2893)
* Allow oblique Mercator maps have a vertical oblique Equator

THrought GMT history, the oblique Equator has always been projected to be the horizontal x-axis.  However, for features with more N-S extend than E-W, it makes more sense to align the plot so that the oblique Equator is the positive y-axis.  This requires a new modifer to -JO (+v) to make it so.

* Rotated view now works

* Add test cases

* Update gmt_map.c

* Update various PS

* Add vertical oblique Equator documentation to cookbook

* Typos

* Update map-projections.rst

* Update GMT_obl_nz.sh

* Change -Rk to -R...+uk in scripts

* Fix error in km scaling

* Update gmt_init.c