Fix grdgdal keys (#2899)
* Fix the KEYS in grdgdal

This module only reads datasets in GMT, but only for two of the methods selectable in -A (grid and rasterize).  For the other methods we have GDAL read grids directly.  Hence the KEYS is just for dataset and if not needed then the primary setting is changed to secondary so that external calls do not result in a forced argument.
Also use GMT_IS_DATASET and GMT_IS_GRID instead of 0 and 1 for readability.

* Update grdgdal.c

* Fix grdgdal keys

Needs ? and set it during runtime