Adjust Mercator projection origin to (lon0, lat0)
The x-cordinates out of the Mercator projection are relative to lon0 but the y-coordinates were always relative to Equator and not lat0.  This caused troubles when relative shifts from an orign was given, such as in #2287.  This PR addresses this issue by keeping j_yc, the y-coordinate of the standard parallel and using that to ensure the origin has coordinates (0,0).  Closes #2287.