doc: Fix CSS style for Sphinx 2.x (#2925)
Our current theme sphinx_rtd_theme v0.4.3 doesn't work well with Sphinx 2.x,
especially for the option lists.

This PR fixes the most common styling issues for Sphinx 2.x and keep
backward compatibility with Sphinx 1.x

Note that this is a temporary workaround. We still need to wait for the
official fix.