Let grdinterpolate -E accept a user profile file (#2927)
* Let grdinterpolate -E accept a user profile file

Instead of having to create the profile from coordinates, allow a single segment with equidistant spacing be allowed as well.  Added new test to make sure it works.

* Let grdinterpolate -E accept a user profile file

Instead of having to create the profile from coordinates, allow a single segment with equidistant spacing be allowed as well.  Added new test to make sure it works.