cmake: Improve cmake script (#2950)
If I understand it correctly, the settings for triangle.c are only needed
when the LICENSE_RESTRICTED is on (i.e. Shewchuk's triangulation is used).
So I move the related codes to the LICENSE_RESTRICTED if-test.

Also move some other codes to better places to make the script more readable.