Let psscale rely on -B machinery for labels if possible (#2952)
* Let psscale rely on -B machiner for labels if possible

The -B machiner is more robust in handling things like label placement provided it is on the default side of the axis, such as the case with #2943.  THe solution was to rely more on that function except for the cases that are different.  Closes #2943. UPdates a few PS originals by a tiny amount.

* Simplify the decision

* Also need to place label if no -B.