Clarify what the Bezier and linear splines do (#2962)
The option in psxy to fit a Bezier spline rather than linear to data points apply to the projected 2-D coordinates, not (say) the original longitude, latitude points on a spherical surface.  I am getting an article on the extension of the Bezier spline to spherical surfaces so possibly this could change in the future.  For Cartesian data this does not matter but it would affect geographic unprojected data.  For now, clearer documentation will do.