Let a signed grid cross size mean centered or asymmetrical grid ticks (#2877)
* Let a negative grid cross size means asymmetrical ticks

See #2875 for context.

* Document purpose negative grid cross sizes

* Add a test

* Allow symmetrical cross-ticks too.

* Implement memory of gridlines setting

For grid ticks one must first lay down a single (primary) set of gridlines, then that spacing sets the line spacing for the denser grid ticks (primary and secondary).  I generalized functions used to set and get current CPT to avoid duplicated code.

* Update gmt.conf.rst

* Simplifications

* Update asymm_grid.sh

* Do not draw ticks at gridlines

* Update asymm_grid.ps

* Enable gridline memory in classic mode

Just writing to a file in the tmp dir and not the session dir, so possibility for clashes between multiple script exists.

* Use enums instead of integer flag for cross-type

* Update gmt_support.c

* Fix wrong direction for N-S ticks and no ticks at or above polar caps

* Add the azimuthal proj test

* Update asymm_grid.ps

* Symmetrical ticks for lon = ±180

* Deal with polar caps and gridticks

* Update script and PS

This uses coarser interval as secondary.  Another PR will deal with sorting primary vs secondary so the -Bs vs -Bp will be less of an issue.

* Automatically make smallest interval the primary interval

if users mix up -Bp and -Bs settings we will swap them.

* Update gmt_init.c

* Flip primary and secondary, only check for non-TIME axes

* Only warn, update docs