Fix seg fault when GMT_DATADIR environment variable is defined but invalid
We've observed instances of gmt seg faulting on startup on some machines but not others. The problem turns out to be users with the GMT_DATADIR environment variable set to nonexistant directories, a case causing the current code to try parsing a null pointer value string. The change detects the case and avoids parsing the null pointer string.