Safeguard against rogue GMT_DATADIR settings (#2974)
As per #2971, if a non-existant path is given vie the environmental variable GMT_DATADIR we might get a SEGV crash (presumably because no checking happens later on).  This PR tests all such directories.  Only lists of directories are not checked since these are used later to look for specific files and will only cause file not found run-time errors.  Closes #2971.