Offset labels and titles if x-annotations are angled (#2975)
* Offset labels and titles if x-annotations are angled

An angled x-annotation means it will take more space in the y-direction. This has consequences for how far labels and titles need to be - yet this was not considered.  We did considered moving the annotations themselves further away depending on the angle, but noting else.

* Update gmt_plot.c

Define a zero-offset slant offset, then if angeled annotations we compute the length of the largest annotation, multiply by sin of the angle and update the slant offset PSL parameter.  This is then use to adjust the y-coordinate of both a label and a plot title for Cartesian frames.

* Update gmt_plot.c

* Use abs value of sine

* Add test