Use more local functions if possible
Some GMT functions are defined in one gmt_*.c file and only used in another.  This means it is exported for no good reason.  Better to physically move the function to the file it is used. This PR simply moves gmt_get_angle from gmt_support.c to gmt_map.c (only place where it is used), then renaming it to map_get_angle and making it GMT_LOCAL (static).