Let plot3d also have the -A option
Same as for plot.  However, for now we require the z-coordinates to be constant, i.e., the line is restricted to lie in the plane. This may be relaxed later but will require the general fixup_path and resample path functions accept (x,y,z) and not (x,y) arrays.  Closes #2997.