Avoid separate file with gridline spacing
The new gridlien tick feature requires us to know the past regular gridline spacings.  Initialy, I implemented this via a separate gridline save file.  However, this meant that every plot that drew gridlines but not gridticks ended up writing a file and then deleting it without every being needed.  This PR places that information in the gmt.history which is always read and written, so new separate file is necessary.