Address shortcomings in grd2kml for global grids
This module grew out of a need to image multibeam bathymetry in Google Earth efficiently, so there were decisions made that made sense for a tiny area but failed when we go 0/360, etc.  Also, minor bug in grdimage for grayshaded images in that it reported back via its status that the plot had all NaNs only.  Finally, changes in psconvert -F did not trickle into calls from grdkml so we ended up with duplicate image extensions.