CI: Cache GMT remote files for testings
Recently, the nightly CI jobs fail on Linux and Windows, but succeed on macOS.

The CI jobs fail because the Linux and Windows agents cann't connect to
any external websites, thus GMT can't download any remote files from the GMT server.
It's not clear if it's a temporary issue of Azure Pipelines.

Currently, the only workaround is using the cache mechanism of Azure Pipelines.
The macOS agents download remote files to ${HOME}/.gmt and save them to
the CI cache, then Linux and Windows agents can restore the cache when
needed.

The only problem is that the Azure Pipelines caches are immutable.
When the remote files are changed, we have to manually update the cache key
to regenerate the caches.