Fix grdimage -A for PPM rasters
The gmtapi_export_image function in GMT which writes PPM grids had a bug in the test for image layout, resulting in nothing being written except for a header.  Also fixes a wring information message from grdimage that implied writing via GDAL when PPM was selected.  Clarify the documentation on PMM and GDAL.