Clean up some GMT src shell scripts
I have moved gmt_make_pattern_include.sh to admin since it is not meant to ever be run again - it just is kept for reference.  The gmt_clip_test.sh is very old and deprecated, so I deleted it.  The gmt_make_enum_dicts.sh gave the wrong name for the API function to get an enum.