Hande font values with equal signs in them
From the olden days, we are allowing the deprecated syntax PAR = value and even PAR=value given to gmt set, despite the syntax being PAR value.  This caused trouble for fonts with outline pens which contains an equal sign.  WHile being bakwards compatible, I removed the [=] from the docs to discourage the use of that old syntax.  We still allow it, but not PAR=value for FONTS.