Add auto-naming mode for working directory (#3270)
It is helpful to run big movie jobs in a temp directory, possibly in /tmp or similar, since these (a) may have more free space and (b) may not be monitored by backup and sync tools.  This PR makes this even simpler by allowing -W (no dir given), which them defaults to /tmp/prefix (i.e., a subdir with the same name as the movie prefix but in the system temp folder.