Mixup in grid x/y array allocation
Unfortunately I made a mistake in implementing the variable x,y arrays.  If these are found we save the non-equidistant arrays in a special array, and then in grd2xyz we use it.  However, if these are not found, gmt will allocate x and y arrays (equidistant) so we should not allocate new ones in grd2xyz.