CI: Always build and deploy docs in master branch (#3278)
We update the GMT documentation frequently, but have to wait for the nightly CI job
to build and view the documentation online.

The CI job for the documentation building takes ~22 minutes. After we migrate all
the animations to YouTube, the time could be decreased to ~10 minutes.

This PR changes the CI settings to rebuild the documentation when the
master branch changes, so that we can view the latest dev documentation
after ~20 minutes, not one night.